Extract function for reusing
